#2635 moved folding settings to PerlFoldingSettings
