+++ plugin/core/src/main/java/com/perl5/lang/perl/idea/folding/PerlFoldingSettings.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2020 Alexandr Evstigneev
+ * Copyright 2015-2023 Alexandr Evstigneev
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,9 +17,20 @@
 package com.perl5.lang.perl.idea.folding;
 
 import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.components.PersistentStateComponent;
+import com.intellij.openapi.components.State;
+import com.intellij.openapi.components.Storage;
+import com.intellij.util.xmlb.XmlSerializerUtil;
+import com.perl5.lang.perl.idea.PerlPathMacros;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 
-public class PerlFoldingSettings {
+@State(
+  name = "PerlCodeFoldingSettings",
+  storages = @Storage(PerlPathMacros.APP_CODEINSIGHT_SETTINGS_FILE)
+)
+public class PerlFoldingSettings implements PersistentStateComponent<PerlFoldingSettings> {
   public boolean COLLAPSE_COMMENTS = true;
   public boolean COLLAPSE_CONSTANT_BLOCKS = false;
   public boolean COLLAPSE_ANON_ARRAYS = false;
@@ -29,6 +40,16 @@ public class PerlFoldingSettings {
   public boolean COLLAPSE_TEMPLATES = false;
   public boolean COLLAPSE_QW = false;
   public boolean COLLAPSE_CHAR_SUBSTITUTIONS = true;
+
+  @Override
+  public @Nullable PerlFoldingSettings getState() {
+    return this;
+  }
+
+  @Override
+  public void loadState(@NotNull PerlFoldingSettings state) {
+    XmlSerializerUtil.copyBean(state, this);
+  }
 
   public static PerlFoldingSettings getInstance() {
     return ApplicationManager.getApplication().getService(PerlFoldingSettings.class);
