+++ plugin/src/main/resources/META-INF/plugin.xml
@@ -265,15 +265,12 @@
     <postStartupActivity implementation="com.perl5.lang.perl.idea.project.Perl5ProjectStartupActivity"/>
     <postStartupActivity implementation="com.perl5.lang.perl.idea.project.PerlProjectDirectoriesConfigurator"/>
 
-    <applicationService serviceInterface="com.perl5.lang.perl.idea.configuration.settings.PerlApplicationSettings"
-                        serviceImplementation="com.perl5.lang.perl.idea.configuration.settings.PerlApplicationSettings"/>
+    <applicationService serviceImplementation="com.perl5.lang.perl.idea.configuration.settings.PerlApplicationSettings"/>
     <applicationService serviceImplementation="com.intellij.openapi.projectRoots.impl.PerlSdkTable"/>
     <additionalLibraryRootsProvider implementation="com.perl5.lang.perl.idea.configuration.settings.sdk.PerlLibraryProvider"/>
     <resolveScopeProvider implementation="com.perl5.lang.perl.idea.configuration.settings.sdk.PerlResolveScopeProvider"/>
-    <applicationService serviceInterface="com.perl5.lang.perl.idea.folding.PerlFoldingSettings"
-                        serviceImplementation="com.perl5.lang.perl.idea.folding.PerlFoldingSettingsImpl"/>
-    <applicationService serviceInterface="com.perl5.lang.perl.idea.codeInsight.Perl5CodeInsightSettings"
-                        serviceImplementation="com.perl5.lang.perl.idea.codeInsight.Perl5CodeInsightSettings"/>
+    <applicationService serviceImplementation="com.perl5.lang.perl.idea.folding.PerlFoldingSettings"/>
+    <applicationService serviceImplementation="com.perl5.lang.perl.idea.codeInsight.Perl5CodeInsightSettings"/>
 
     <extendWordSelectionHandler implementation="com.perl5.lang.perl.idea.editor.PerlVariableSelectionHandler"/>
     <basicWordSelectionFilter implementation="com.perl5.lang.perl.idea.editor.PerlVariableSelectionHandler$OuterElementsSuppressor"/>
